requirements.txt
customtkinter
pillow
python-chess
chess